KristianRosenvold: Issue 923 setting socket reusable
This avoids the TIME_WAIT isue, and works well. Please note
that the empheral socket pool is still extremely small on windows, and
may have to be adjusted unless we do further tweaks of the
socket usage elsewhere. We avoid exhausting 30000 sockets on linux, though ;)

The fix for this patch is based on
http://wiki.apache.org/HttpComponents/FrequentlyAskedConnectionManagementQuestions

git-svn-id: http://selenium.googlecode.com/svn/trunk@10082 07704840-8298-11de-bf8c-fd130f914ac9