JasonLeyba: When webdriver.element.getText is compiled as a js_fragment, the "//*[@usemap]" XPath expression in bot.dom.isShown confuses the Closure compiler, which incorrectly identifies the start of a comment block. This results in invalid JS being generated since a large portion of code is stripped out.  I haven't been able figure out what the compiler is matching the "/*" against to close the block, but switching to a "/descendant::*" selector instead of "//*" fixes the bug.
I've added a TODO to remind myself to try to find a reproducible case for //* so we can file a bug against the closure-compiler project.  Also, the useMap attribute is only valid on a small subset of elements, so I've added a TODO to further refine the selector to only consider those elements.


git-svn-id: http://selenium.googlecode.com/svn/trunk@10756 07704840-8298-11de-bf8c-fd130f914ac9