KristianRosenvold: Fixed persistent connections in java client.
Tested with selenium-server and connection re-use works fine.
This means it's now possible to run a central build server
with distributed remote servers without running out of
ephemeral sockets.

This may also fix the problem for some browsers, but I have not
tested this. It does not fix firefox running on the same machine
as FirefoxDriver.

git-svn-id: http://selenium.googlecode.com/svn/trunk@11199 07704840-8298-11de-bf8c-fd130f914ac9