StuartKnightley: Add "nice" forms of strategy names
(14:12:23) jarib: atm we send 'class name' across the wire
(14:12:27) jarib: IIRC
(14:12:28) simonstewart: Basically, we use the translated by from your
example to look up a property of the lcoators map
(14:12:34) simonstewart: Right, and that's lovely
(14:12:38) simonstewart: we should keep doing that
(14:12:48) simonstewart: Just make the atoms aware of the alternative
names
(14:12:58) ato: Right.
(14:12:59) ato: jarib: Indeed.
(14:13:01) jarib: so the translation doesn't happen in the driver.
(14:13:16) ato: I suppose there's no chance of a name collision in the
future here.
(14:13:38) simonstewart: It happens in the atoms
(14:13:46) simonstewart: Gimme a sec
(14:14:16) simonstewart:
http://code.google.com/p/selenium/source/browse/trunk/javascript/atoms/locators/locators.js#45
(14:14:27) simonstewart: We just need to shove some values into that

git-svn-id: http://selenium.googlecode.com/svn/trunk@13745 07704840-8298-11de-bf8c-fd130f914ac9