KristianRosenvold: Changed all the fxtimers to be persisted at a guaranteed gc-safe scope
Prior to this fix things were stable on linux, but repeated experiments
on windows XP seem to indicate that javascript gc is a great deal smarter than we
think,- and it seems to be platform specific. On XP it would appear to easily spot that a
circular reference between the timer and its closure can be gc'ed, (which is correct in
terms of GC but not in the baheviour we want), so someone else needs to be holding
on to the timer.

I am not happy with the overall levels of scientific correctness in this fix,
as I somehow feel I am just rearranging deck-chairs to make some other problem
go away. Fortunately, it's not the titanic, so moving the deckchairs won't hurt.

Pretty it is not (Yoda)

git-svn-id: http://selenium.googlecode.com/svn/trunk@16301 07704840-8298-11de-bf8c-fd130f914ac9