JimEvans: Allows the user to disable native events in favor of synthetic (JavaScript) events in the IE driver. This would alleviate the problems with, for example, hovering over elements with the mouse to access a JavaScript-based menu framework. Note that using synthetic events will cause user code to block if an JavaScript alert is displayed while performing the event. This will be a limitation of synthetic events, and will not be fixed for the foreseeable future.
While this commit does provide a part of this functionality, it is not complete, so it should not be used yet. Drag-and-drop of elements is still plagued by the same mouse cursor position problem as native events, wherein the drag is affected by the physical mouse cursor being within the IE window bounds, and as such, synthetic events in IE should not be used yet. Drag and drop will not work for complex (e.g., jQuery) elements in IE9, so this functionality should not be used yet. Keyboard events do not respect non printable keys like shift, control, alt, home, delete, and the like, so synthetic events in IE should not be used yet. Finally, to reiterate a central theme of this commit, SYNTHETIC EVENTS IN IE SHOULD NOT BE USED YET.

git-svn-id: http://selenium.googlecode.com/svn/trunk@16888 07704840-8298-11de-bf8c-fd130f914ac9